+++ src/commands/build.js
@@ -11,15 +11,15 @@ class TriggerBuildCommand extends Command {
    */
   async run() {
     const { flags, args } = this.parse(TriggerBuildCommand);
-    const { env } = args; // package is reserved so we don't try to destructure it
-    const { pkg, version } = this.parsePackage(args.package);
+    const { env, package: packageArg } = args;
+    const { pkg, version } = this.parsePackage(packageArg);
     const { wrhsHost, auth } = this.mergeConfig(flags);
 
     if (!wrhsHost) {
-      this.error(this.missingHostError());
+      return this.error(this.missingHostError());
     }
     if (!version) {
-      this.error(this.missingVersionError());
+      return this.error(this.missingVersionError('build'));
     }
 
     const wrhs = this.wrhs(auth, { wrhsHost });
@@ -28,7 +28,7 @@ class TriggerBuildCommand extends Command {
     return new Promise((resolve, reject) => {
       wrhs.builds.trigger({ env, pkg, version, promote: flags.promote }, (err, response) => {
         if (err) {
-          this.renderError('build', args.package, err);
+          this.renderError('build', packageArg, err);
           return reject(err);
         }
         if (flags.json) {