+++ test/commands/promote.test.js
@@ -13,8 +13,8 @@ const mockConfig = {
 
 const generateMockWarehouseRoute = (options = {}) => {
   const {
-    path = '/builds/package/dev/version',
-    query = { promote: false },
+    path = '/promote/package/dev/version',
+    query = { build: false },
     auth = mockConfig.auth
   } = options;
 
@@ -28,10 +28,15 @@ const generateMockWarehouseRoute = (options = {}) => {
 };
 
 const validate = ({ stdout }) => {
-  assume(stdout).contains(`Build triggered successfully. You can use 'wrhs get:status package@version dev' to get the status of this build`);
+  assume(stdout).contains(`Promote triggered successfully. You can use 'wrhs get:build package dev' to confirm that your specific version has been promoted.`);
 };
 
-describe('build', function () {
+const validateWithBuild = ({ stdout }) => {
+  assume(stdout).contains(`Build triggered successfully. You can use 'wrhs get:status package@version dev' to get the status of this build`);
+  assume(stdout).contains(`Promote will be triggered upon successful build. You can use 'wrhs get:build package dev' to confirm that your specific version has been promoted.`);
+};
+
+describe('promote', function () {
   before(function () {
     sinon.stub(fs, 'readFileSync')
       .withArgs(sinon.match('.wrhs'), 'utf8')
@@ -44,20 +49,20 @@ describe('build', function () {
     sinon.restore();
   });
 
-  // Trigger a build with specific version
+  // Can promote with specific version
   test
     .nock('https://warehouse.ai', generateMockWarehouseRoute())
     .stdout()
-    .command(['build', 'package@version', 'dev'])
-    .it('can trigger build with specific version', validate);
+    .command(['promote', 'package@version', 'dev'])
+    .it('can promote with specific version', validate);
 
-  // Fails to trigger a build without specific version supplied
+  // Fails to promote without specific version supplied
   test
     .stdout()
-    .command(['build', 'package', 'dev'])
+    .command(['promote', 'package', 'dev'])
     .catch(err => {
       assume(err.oclif.exit).equals(2);
-      assume(err.message).contains('Missing package version. Please sure package is in the form `packageName@version` where `version` is the specific version to build');
+      assume(err.message).contains('Missing package version. Please make sure `package` is in the form `packageName@version` where `version` is the specific version to promote');
     })
     .it('Outputs an error if there is no version supplied as part of package');
 
@@ -67,8 +72,8 @@ describe('build', function () {
       auth: { user: 'userFlag', pass: 'passFlag' }
     }))
     .stdout()
-    .command(['build', 'package@version', 'dev', '-u', 'userFlag', '-p', 'passFlag'])
-    .it('can trigger build with passed in auth using abbreviated flag names', validate);
+    .command(['promote', 'package@version', 'dev', '-u', 'userFlag', '-p', 'passFlag'])
+    .it('can promote with passed in auth using abbreviated flag names', validate);
 
   // -user and -pass flags
   test
@@ -76,24 +81,24 @@ describe('build', function () {
       auth: { user: 'userFlag', pass: 'passFlag' }
     }))
     .stdout()
-    .command(['build', 'package@version', 'dev', '--user', 'userFlag', '--pass', 'passFlag'])
-    .it('can trigger build with passed in auth using full flag names', validate);
+    .command(['promote', 'package@version', 'dev', '--user', 'userFlag', '--pass', 'passFlag'])
+    .it('can promote with passed in auth using full flag names', validate);
 
   // -h flag
   test
     .nock('https://wrhs.passedinHost', generateMockWarehouseRoute())
     .stdout()
-    .command(['build', 'package@version', 'dev', '-h', 'wrhs.passedinHost'])
-    .it('can trigger a build from a passed in host', validate);
+    .command(['promote', 'package@version', 'dev', '-h', 'wrhs.passedinHost'])
+    .it('can promote from a passed in host', validate);
 
-  // -m flag
+  // -b flag
   test
     .nock('https://warehouse.ai', generateMockWarehouseRoute({
-      query: { promote: true }
+      query: { build: true }
     }))
     .stdout()
-    .command(['build', 'package@version', 'dev', '-m'])
-    .it('can trigger a build with promote', validate);
+    .command(['promote', 'package@version', 'dev', '-b'])
+    .it('can build and promote', validateWithBuild);
 
   // Host fallback
   test
@@ -110,6 +115,6 @@ describe('build', function () {
     })
     .nock('https://wrhs-host-fallback.ai', generateMockWarehouseRoute())
     .stdout()
-    .command(['build', 'package@version', 'dev'])
+    .command(['promote', 'package@version', 'dev'])
     .it('fallsback to the `host` config', validate);
 });